+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 3/8/26:
+- Exposed transcoding settings on Settings > Library page, and option to re-transcode for new scan.
 - "Delete from Disk" Movie Details page option added.
 - Fixes to normal+mini player, align styles to the same.
 - Stats broken out in About section, more details, hover info.